Crate magnus

Magnus is a library for writing Ruby extentions in Rust, or running Ruby code from Rust.

Overview

All Ruby objects are represented by Value. To make it easier to work with values that are instances of specific classes a number of wrapper types are available. These wrappers will Deref to Value, so you can still use Value’s methods on them.

Ruby Class	Magnus Type

String	RString

Integer	Integer

Float	Float

Array	RArray

Hash	RHash

Symbol	Symbol

Class	RClass

Module	RModule

When writing Rust code to be called from Ruby the init attribute can be used to mark your init function that Ruby will call when your library is required.

When embedding Ruby in a Rust program, see [embed::init] for initialising the Ruby VM.

The method macro can be used to wrap a Rust function with automatic type conversion and error handing so it can be exposed to Ruby. The TryConvert trait handles conversions from Ruby to Rust, and anything implementing IntoValue can be returned to Ruby. See the Module and Object traits for defining methods.

Value::funcall can be used to call Ruby methods from Rust.

See the wrap attribute macro for wrapping Rust types as Ruby objects.

Safety

When using Magnus, in Rust code, Ruby objects must be kept on the stack. If objects are moved to the heap the Ruby GC can not reach them, and they may be garbage collected. This could lead to memory safety issues.

It is not possible to enforce this rule in Rust’s type system or via the borrow checker, users of Magnus must maintain this rule manually.

An example of something that breaks this rule would be storing a Ruby object in a Rust heap allocated data structure, such as Vec, HashMap, or Box. This must be avoided at all costs.

While it would be possible to mark any functions that could expose this unsafty as unsafe, that would mean that almost every interaction with Ruby would be unsafe. This would leave no way to differentiate the really unsafe functions that need much more care to use.

Examples

use magnus::{class, define_module, function, method, prelude::*, Error};

#[magnus::wrap(class = "Euclid::Point", free_immediately, size)]
struct Point {
    x: isize,
    y: isize,
}

impl Point {
    fn new(x: isize, y: isize) -> Self {
        Self { x, y }
    }

    fn x(&self) -> isize {
        self.x
    }

    fn y(&self) -> isize {
        self.y
    }
}

fn distance(a: &Point, b: &Point) -> f64 {
    (((b.x - a.x).pow(2) + (b.y - a.y).pow(2)) as f64).sqrt()
}

#[magnus::init]
fn init() -> Result<(), Error> {
    let module = define_module("Euclid")?;
    let class = module.define_class("Point", class::object())?;
    class.define_singleton_method("new", function!(Point::new, 2))?;
    class.define_method("x", method!(Point::x, 0))?;
    class.define_method("y", method!(Point::y, 0))?;
    module.define_module_function("distance", function!(distance, 2))?;
    Ok(())
}

Crates that work with Magnus

rb-sys - low level bindings to Ruby.

serde_magnus - Serde integration.

C Function Index

Click to show

This lists all the Ruby C API functions currently implemented, and how they map to Rust functions and methods in Magnus.

A-N

Data_Get_Struct: See wrap and TypedData.

Data_Make_Struct: See wrap and TypedData.

Data_Wrap_Struct: See wrap and TypedData.

RARRAY

RARRAY: Similar to RArray::from_value.

RARRAY_ASET: Similar to RArray::store.

RARRAY_CONST_PTR: Similar to RArray::as_slice.

RARRAY_EMBED_LEN: See RArray::len.

RARRAY_LEN: RArray::len.

RARRAY_LENINT: RArray::len.

RARRAY_PTR: Similar to RArray::as_slice.

RB

RBIGNUM_NEGATIVE_P: See RBignum::is_negative.

RBIGNUM_POSITIVE_P: See RBignum::is_positive.

rb_a-rb_arx

rb_alias: Module::define_alias.

rb_any_to_s: std::fmt::Display.

rb_ary

rb_ary_assoc: RArray::assoc.

rb_ary_cat: RArray::cat.

rb_ary_clear: RArray::clear.

rb_ary_cmp: RArray::cmp.

rb_ary_concat: RArray::concat.

rb_ary_delete: RArray::delete.

rb_ary_delete_at: RArray::delete_at.

rb_ary_dup: Similar to RArray::dup.

rb_ary_each: See RArray::each.

rb_ary_entry: RArray::entry.

rb_ary_freeze: See Value::freeze.

rb_ary_includes: RArray::includes.

rb_ary_new: RArray::new.

rb_ary_new_capa: RArray::with_capacity.

rb_ary_new_from_args: Not implemented, see RArray::from_slice.

rb_ary_new_from_values: RArray::from_slice.

rb_ary_plus: RArray::plus.

rb_ary_pop: RArray::pop.

rb_ary_push: RArray::push.

rb_ary_rassoc: RArray::rassoc.

rb_ary_replace: RArray::replace.

rb_ary_resize: RArray::resize.

rb_ary_reverse: RArray::reverse.

rb_ary_rotate: RArray::rotate.

rb_ary_shared_with_p: RArray::is_shared.

rb_ary_shift: RArray::shift.

rb_ary_sort: Not implemented, see RArray::sort.

rb_ary_sort_bang: RArray::sort.

rb_ary_store: RArray::store.

rb_ary_subseq: RArray::subseq.

rb_ary_to_ary: RArray::to_ary.

rb_ary_to_s: Not implemented directly, see Value::to_r_string.

rb_ary_unshift: RArray::unshift.

rb_as-rb_az

rb_ascii8bit_encindex: encoding::Index::ascii8bit.

rb_ascii8bit_encoding: RbEncoding::ascii8bit.

rb_attr: Module::define_attr.

rb_b

rb_backref_get: backref_get.

rb_binding_new: [Binding::new]. Unimplemented >= Ruby 3.2.

rb_block_call: Value::block_call.

rb_block_given_p: block::block_given.

rb_block_proc: block::block_proc.

rb_bug: error::bug.

rb_c

rb_call_super: call_super.

rb_check_arity: scan_args::check_arity.

rb_check_array_type: See TryConvert and Value::try_convert.

rb_check_funcall: Value::check_funcall.

rb_check_hash_type: See TryConvert and Value::try_convert.

rb_check_id: Similar to Id::check.

rb_check_id_cstr: Id::check.

rb_check_symbol: Similar to StaticSymbol::check.

rb_check_symbol_cstr: StaticSymbol::check.

rb_check_typeddata: See TryConvert and Value::try_convert.

rb_class2name: RClass::name.

rb_class_name: Simmilar to Value::classname.

rb_class_new: RClass::new.

rb_class_new_instance: RClass::new_instance.

rb_class_superclass: RClass::superclass.

rb_complex_abs: RComplex::abs.

rb_complex_arg: RComplex::arg.

rb_complex_conjugate: RComplex::conjugate.

rb_complex_imag: RComplex::imag.

rb_complex_new: RComplex::new.

rb_complex_new_polar: RComplex::polar.

rb_complex_real: RComplex::real.

rb_const_get: Module::const_get.

rb_const_set: Module::const_set.

rb_current_receiver: current_receiver.

rb_d

rb_data_object_make: See wrap and TypedData.

rb_data_object_wrap: See wrap and TypedData.

rb_data_object_zalloc: See wrap and TypedData.

rb_data_typed_object_make: See wrap and TypedData.

rb_data_typed_object_wrap: See wrap and TypedData.

rb_data_typed_object_zalloc: See wrap and TypedData.

rb_default_external_encoding: RbEncoding::default_external.

rb_default_internal_encoding: RbEncoding::default_internal.

rb_define_alias: Module::define_alias.

rb_define_attr: See Module::define_attr.

rb_define_class: define_class.

rb_define_class_id: Simmilar to define_class.

rb_define_class_id_under: Module::define_class.

rb_define_class_under: See Module::define_class.

rb_define_global_function: define_global_function.

rb_define_method: See Module::define_method.

rb_define_method_id: Module::define_method.

rb_define_module: define_module.

rb_define_module_function: RModule::define_module_function.

rb_define_module_id: See define_module.

rb_define_module_id_under: Module::define_module.

rb_define_module_under: See Module::define_module.

rb_define_private_method: Module::define_private_method.

rb_define_protected_method: Module::define_protected_method.

rb_define_singleton_method: Object::define_singleton_method.

rb_define_variable: define_variable.

rb_e-rb_enb

rb_enc

rb_enc_ascget: RbEncoding::ascget.

rb_enc_associate: See EncodingCapable::enc_associate.

rb_enc_associate_index: EncodingCapable::enc_associate.

rb_enc_check: encoding::check.

rb_enc_codelen: RbEncoding::codelen.

rb_enc_codepoint_len: RbEncoding::codepoint_len.

RB_ENC_CODERANGE: RString::enc_coderange.

RB_ENC_CODERANGE_CLEAR: RString::enc_coderange_clear.

RB_ENC_CODERANGE_SET: RString::enc_coderange_set.

rb_enc_compatible: encoding::compatible.

rb_enc_copy: encoding::copy.

rb_enc_default_external: Encoding::default_external. Encoding::default_internal.

rb_enc_fast_mbclen: RbEncoding::fast_mbclen.

rb_enc_find: RbEncoding::find.

rb_enc_find_index: encoding::Index::find.

rb_enc_from_encoding: std::convert::From.

rb_enc_from_index: std::convert::From.

rb_enc_get: Use EncodingCapable::enc_get plus std::convert::From.

rb_enc_get_index: EncodingCapable::enc_get.

rb_enc_mbclen: RbEncoding::mbclen.

rb_enc_precise_mbclen: RbEncoding::precise_mbclen.

rb_enc_reg_new: RRegexp::new.

rb_enc_set_index: EncodingCapable::enc_set.

rb_enc_str_coderange: RString::enc_coderange_scan.

rb_enc_str_new: RString::enc_new.

rb_enc_to_index: std::convert::From.

rb_enc_uint_chr: RbEncoding::chr.

rb_en-rb_ez

rb_enumeratorize: See Value::enumeratorize.

rb_enumeratorize_with_size: Simmilar to Value::enumeratorize.

rb_eql: Value::eql.

rb_equal: Value::equal.

rb_error_arity: scan_args::check_arity.

rb_eval_string: See eval() or eval!.

rb_eval_string_protect: eval() or eval!.

rb_exc_raise: Return Error.

rb_extend_object: Object::extend_object.

rb_f

rb_filesystem_encindex: encoding::Index::filesystem.

rb_filesystem_encoding: RbEncoding::filesystem.

rb_find_encoding: std::convert::From.

rb_float_new: RFloat::from_f64 or Float::from_f64.

rb_float_new_in_heap: See Float::from_f64.

rb_flt_rationalize: Float::rationalize.

rb_flt_rationalize_with_prec: Float::rationalize_with_prec.

rb_funcall: See Value::funcall.

rb_funcallv: Value::funcall.

rb_funcallv_public: Value::funcall_public.

rb_funcall_with_block: Value::funcall_with_block.

rb_g

rb_gc: gc::start.

rb_gc_adjust_memory_usage: gc::adjust_memory_usage.

rb_gc_count: gc::count.

rb_gc_disable: gc::disable.

rb_gc_enable: gc::enable.

rb_gc_location: gc::location.

rb_gc_mark: gc::mark.

rb_gc_mark_locations: gc::mark_slice.

rb_gc_mark_movable: gc::mark_movable.

rb_gc_register_address: gc::register_address or BoxValue.

rb_gc_register_mark_object: gc::register_mark_object.

rb_gc_start: gc::start.

rb_gc_stat: gc::stat or gc::all_stats.

rb_gc_unregister_address: gc::unregister_address.

rb_get_kwargs: scan_args::get_kwargs.

rb_get_path: TryConvert/Value::try_convert to std::path::PathBuf.

rb_h

rb_hash: Value::hash.

rb_hash_aref: RHash::aref.

rb_hash_aset: RHash::aset.

rb_hash_bulk_insert: RHash::bulk_insert.

rb_hash_clear: RHash::clear.

rb_hash_delete: RHash::delete.

rb_hash_fetch: RHash::fetch.

rb_hash_foreach: RHash::foreach.

rb_hash_lookup: RHash::lookup.

rb_hash_lookup2: RHash::get.

rb_hash_new: RHash::new.

rb_hash_new_capa: RHash::with_capacity.

rb_hash_size: RHash::size.

rb_hash_size_num: RHash::len.

rb_hash_update_by: RHash::update (update_func arg not implemented).

rb_i-rb_in

rb_id2name: Id::name.

rb_id2sym: std::convert::From.

rb_include_module: Module::include_module.

rb_inspect: Value::inspect or std::fmt::Debug.

rb_intern: std::convert::From.

rb_intern2: std::convert::From.

rb_intern3: std::convert::From.

rb_intern_str: std::convert::From.

rb_io

rb_is-rb_iz

rb_ivar_get: Object::ivar_get.

rb_ivar_set: Object::ivar_set.

rb_j-rb_k

rb_jump_tag: Return Error.

rb_l

rb_locale_encindex: encoding::Index::locale.

rb_locale_encoding: RbEncoding::locale.

rb_m

rb_module_new: RModule::new.

rb_mod_ancestors: Module::ancestors.

rb_n

rb_num_coerce_bin: Numeric::coerce_bin.

rb_num_coerce_bit: Numeric::coerce_bit.

rb_num_coerce_cmp: Numeric::coerce_cmp.

rb_num_coerce_relop: Numeric::coerce_relop.

rb_o

rb_obj_as_string: Value::to_r_string.

rb_obj_classname: Value::classname.

rb_obj_freeze: Value::freeze.

rb_obj_is_kind_of: Value::is_kind_of.

rb_obj_is_proc: Proc::from_value.

rb_obj_respond_to: Value::respond_to.

rb_p

rb_proc_arity: Proc::arity.

rb_proc_call: Proc::call.

rb_proc_lambda_p: Proc::is_lambda.

rb_proc_new: Proc::new & Proc::from_fn.

rb_protect: Called internally by Magnus when required. Available as [rb_sys::protect] with rb-sys-interop feature for calling raw Ruby api.

rb_r

rb_raise: Simmilar to returning Error.

rb_range_beg_len: Range::beg_len.

rb_rational_den: RRational::den.

rb_rational_new: RRational::new.

rb_rational_num: RRational::num.

rb_reg_backref_number: RMatch::backref_number.

rb_reg_last_match: RMatch::matched.

rb_reg_match: RRegexp::reg_match.

rb_reg_match_last: RMatch::last.

rb_reg_match_post: RMatch::post.

rb_reg_match_pre: RMatch::pre.

rb_reg_new: See RRegexp::new.

rb_reg_new_str: RRegexp::new_str.

rb_reg_nth_defined: RMatch::nth_defined.

rb_reg_nth_match: RMatch::nth_match.

rb_reg_options: RRegexp::options.

rb_require: require.

rb_require_string: require.

rb_s-rb_strl

rb_scan_args: scan_args::scan_args.

rb_singleton_class: Object::singleton_class.

rb_struct

rb_struct_aref: RStruct::aref.

rb_struct_aset: RStruct::aset.

rb_struct_define: r_struct::define_struct.

rb_struct_getmember: RStruct::getmember.

rb_struct_members: RStruct::members.

rb_struct_new: See RClass::new_instance.

rb_struct_size: RStruct::size.

rb_str

rb_str_buf_append: RString::buf_append.

rb_str_buf_cat: RString::cat.

rb_str_buf_cat_ascii: See RString::cat.

rb_str_buf_new: RString::buf_new.

rb_str_buf_new_cstr: See RString::buf_new + RString::cat.

rb_str_capacity: RString::capacity.

rb_str_cat: RString::cat.

rb_str_cat_cstr: See RString::cat.

rb_str_cmp: RString::cmp.

rb_str_conv_enc: RString::conv_enc.

rb_str_drop_bytes: RString::drop_bytes.

rb_str_dump: RString::dump.

rb_str_dup_frozen: See RString::new_frozen.

rb_str_ellipsize: RString::ellipsize.

rb_str_new: RString::from_slice.

rb_str_new_frozen: RString::new_frozen.

rb_str_new_lit: Simmilar to r_string!.

rb_str_new_literal: Simmilar to r_string!.

rb_str_new_shared: RString::new_shared.

rb_str_offset: RString::offset.

rb_str_plus: RString::plus.

rb_str_replace: RString::replace.

rb_str_scrub: RString::scrub.

rb_str_shared_replace: RString::shared_replace.

rb_str_split: RString::split.

rb_str_strlen: RString::length.

rb_str_times: RString::times.

rb_str_to_interned_str: RString::to_interned_str.

rb_str_to_str: TryConvert or Value::try_convert.

rb_str_update: RString::update.

rb_st_

rb_sy-rb_sz

rb_sym2id: std::convert::From.

rb_sym2str: Symbol::name.

rb_t

RB_TEST: Value::to_bool / TryConvert / Value::try_convert.

rb_to_encoding: TryConvert or Value::try_convert.

rb_to_encoding_index: TryConvert or Value::try_convert.

rb_to_float: TryConvert or Value::try_convert.

rb_to_int: TryConvert or Value::try_convert.

rb_to_symbol: std::convert::From.

rb_u

rb_undef_alloc_func: Class::undef_alloc_func

rb_usascii_encindex: encoding::Index::usascii.

rb_usascii_encoding: RbEncoding::usascii.

rb_utf8_encindex: encoding::Index::utf8.

rb_utf8_encoding: RbEncoding::utf8.

rb_utf8_str_new: RString::new.

rb_utf8_str_new_cstr: See RString::new.

rb_utf8_str_new_lit: Simmilar to r_string!.

rb_utf8_str_new_literal: Simmilar to r_string!.

rb_utf8_str_new_static: r_string!.

rb_v-rb_z

rb_warning: error::warning.

rb_yield: block::yield_value / return block::Yield.

rb_yield_splat: block::yield_splat / return block::YieldSplat.

rb_yield_values: See block::yield_values / return block::YieldValues.

rb_yield_values2: block::yield_values / return block::YieldValues.

rc-rt

RSTRING: Similar to RString::from_value.

RSTRING_EMBED_LEN: Similar to RString::len.

RSTRING_LEN: RString::len.

RSTRING_LENINT: RString::len.

RSTRING_PTR: Similar to RString::as_str and RString::as_slice.

RTEST: Value::to_bool / TryConvert / Value::try_convert.

ruby_

ruby_cleanup: See [embed::init] and [embed::Cleanup].

ruby_script: Similar to [embed::ruby_script].

ruby_setup: [embed::setup].

ruby_set_script_name: [embed::ruby_script].

S-Z

TypedData_Get_Struct: See wrap and TypedData.

TypedData_Make_Struct: See wrap and TypedData.

TypedData_Wrap_Struct: See wrap and TypedData.

Re-exports

pub use crate::class::Class;

pub use crate::class::RClass;

pub use crate::error::Error;

pub use crate::exception::Exception;

pub use crate::exception::ExceptionClass;

pub use crate::module::Attr;

pub use crate::module::Module;

pub use crate::module::RModule;

pub use crate::numeric::Numeric;

pub use crate::r_hash::RHash;

pub use crate::r_regexp::RRegexp;

pub use crate::r_string::RString;

pub use crate::r_struct::RStruct;

pub use crate::typed_data::DataType;

pub use crate::typed_data::DataTypeFunctions;

pub use crate::typed_data::TypedData;

pub use crate::value::Fixnum;

pub use crate::value::StaticSymbol;

pub use crate::value::Value;

pub use crate::value::QFALSE;

pub use crate::value::QNIL;

pub use crate::value::QTRUE;

Modules

block

Types and functions for working with Ruby blocks and Procs.

class

Types and functions for working with Ruby classes.

encoding

Types and functions for working with encodings.

error

Rust types for working with Ruby Exceptions and other interrupts.

exception

Types and functions for working with Ruby exceptions.

gc

Functions for working with Ruby’s Garbage Collector.

method

Types and Traits for exposing Rust functions as Ruby methods.

module

Types and functions for working with Ruby modules.

numeric

Types and Traits for working with Ruby’s Numeric class.

prelude

Traits that commonly should be in scope.

r_hash

Types and functions for working with Ruby’s Hash class.

r_regexp

Types for working with Ruby’s Regexp class.

r_string

Types for working with Ruby’s String class.

r_struct

Types and functions for working with Ruby’s Struct class.

scan_args

Types and functions for complex method arguments.

typed_data

Types and Traits for wrapping Rust types as Ruby objects.

value

Types for working with Ruby’s VALUE type, representing all objects, and ‘immediate’ values such as Fixnum.

Macros

eval

Evaluate a literal string of Ruby code with the given local variables.

function

Wrap a Rust function item with Ruby type conversion and error handling, ignoring Ruby’s self argument.

memoize

Utility to simplify initialising a static with std::sync::Once.

method

Wrap a Rust function item with Ruby type conversion and error handling.

r_string

Create a RString from a Rust str literal.

Structs

Binding

A Value known to be an instance of Binding.

Enumerator

Wrapper type for a Value known to be an instance of Ruby’s Enumerator class.

Float

A type wrapping either a Flonum value or a Value known to be an instance of Float.

Flonum

A Value known to be a flonum, Ruby’s internal representation of lower precision floating point numbers.

Integer

A type wrapping either a Fixnum or a RBignum.

RArray

A Value pointer to a RArray struct, Ruby’s internal representation of an Array.

RBignum

A Value pointer to a RBignum struct, Ruby’s internal representation of large integers.

RComplex

A Value pointer to a RComplex struct, Ruby’s internal representation of complex numbers.

RFile

A Value pointer to a RFile struct, Ruby’s internal representation of files.

RFloat

A Value pointer to a RFloat struct, Ruby’s internal representation of high precision floating point numbers.

RMatch

A Value pointer to a RMatch struct, Ruby’s internal representation of the MatchData returned from a regex match.

RObject

A Value pointer to a RObject struct, Ruby’s internal representation of generic objects, not covered by the other R* types.

RRational

A Value pointer to a RRational struct, Ruby’s internal representation of rational numbers.

RTypedData

A Value pointer to a RTypedData struct, Ruby’s internal representation of objects that wrap foreign types.

Range

Wrapper type for a Value known to be an instance of Ruby’s Range class.

Symbol

A type wrapping either a StaticSymbol or a Value pointer to a RSymbol struct.

Traits

ArgList

Trait for types that can be used as an arguments list when calling Ruby methods.

IntoValue

Conversions from Rust types into Value.

IntoValueFromNative

Conversions from Rust types that do not contain Value into Value.

Object

Functions available all non-immediate values.

TryConvert

Conversions from Value to Rust types.

Functions

backref_get

Returns the result of the most recent regexp match.

call_super

Call the super method of the current method context.

current_receiver

Return the Ruby self of the current method context.

define_class

Define a class in the root scope.

define_error

Define an exception class in the root scope.

define_global_const

Define a global constant.

define_global_function

Define a method in the root scope.

define_module

Define a module in the root scope.

define_variable

Define a global variable.

eval

Evaluate a string of Ruby code, converting the result to a T.

require

Finds and loads the given feature if not already loaded.

Attribute Macros

init

Mark a function as the ‘init’ function to be run for a library when it is required by Ruby code.

wrap

Allow a Rust type to be passed to Ruby, automatically wrapped as a Ruby object.

Derive Macros

DataTypeFunctions

Derives DataTypeFunctions with default implementations, for simple uses of TypedData.

TypedData

Derives TypedData, allowing the type to be passed to Ruby automatically wrapped as a Ruby object.