Struct magnus::class::RClass

#[repr(transparent)]
pub struct RClass(_);

A Value pointer to a RClass struct, Ruby’s internal representation of classes.

See the Module trait for defining instance methods and nested classes/modules. See the Object trait for defining singlton methods (aka class methods).

All Value methods should be available on this type through Deref, but some may be missed by this documentation.

Implementations

impl RClass

pub fn from_value(val: Value) -> Option<Self>

Return Some(RClass) if val is a RClass, None otherwise.

Examples

use magnus::{eval, RClass};

assert!(RClass::from_value(eval("String").unwrap()).is_some());
assert!(RClass::from_value(eval("Enumerable").unwrap()).is_none());
assert!(RClass::from_value(eval("nil").unwrap()).is_none());

pub fn new(superclass: RClass) -> Result<RClass, Error>

Create a new anonymous class.

Examples

use magnus::{class, RClass};

let class = RClass::new(class::object()).unwrap();
assert!(class.is_kind_of(class::class()));

pub fn new_instance<T>(self, args: T) -> Result<Value, Error>where
T: ArgList,

Create a new object, an instance of self, passing the arguments args to the initialiser.

Examples

use magnus::class;

let s = class::string().new_instance(()).unwrap();
assert!(s.is_kind_of(class::string()));
assert_eq!(s.to_string(), "");

pub fn superclass(self) -> Result<Self, Error>

Returns the parent class of self.

Returns Err if self can not have a parent class.

Examples

use magnus::{class, eval};

let klass = class::hash().superclass().unwrap();
assert!(klass.equal(class::object()).unwrap());

pub unsafe fn name(&self) -> Cow<'_, str>

Return the name of self.

Safety

Ruby may modify or free the memory backing the returned str, the caller must ensure this does not happen.

This can be used safely by immediately calling into_owned on the return value.

Examples

use magnus::{class, eval};

let value = class::hash();
// safe as we neve give Ruby a chance to free the string.
let s = unsafe { value.name() }.into_owned();
assert_eq!(s, "Hash");