Struct magnus::encoding::Index

#[repr(transparent)]
pub struct Index(_);

The index of an encoding in Ruby’s internal encodings table.

This is the type Ruby uses to label encoding capable types, so is used with operations that require reading or setting that label.

Implementations

impl Index

pub fn ascii8bit() -> Self

Returns the index for ASCII-8BIT a.k.a. binary.

Panics

Panics if called from a non-Ruby thread.

pub fn utf8() -> Self

Returns the index for UTF-8.

Panics

Panics if called from a non-Ruby thread.

pub fn usascii() -> Self

Returns the index for US-ASCII.

Panics

Panics if called from a non-Ruby thread.

pub fn locale() -> Self

Returns the index for the process’ current locale encoding.

This is dynamic. If you change the process’ locale that should also change the return value of this function.

Panics

Panics if called from a non-Ruby thread.

pub fn filesystem() -> Self

Returns the index for filesystem encoding.

This is the encoding that Ruby expects data from the OS’ file system to be encoded as, such as directory names.

Panics

Panics if called from a non-Ruby thread.

pub fn find(name: &str) -> Result<Self, Error>

Returns the index for the encoding with the name or alias name.

Panics

Panics if called from a non-Ruby thread.

Examples

use magnus::{eval, encoding};

assert!(encoding::Index::find("UTF-8").is_ok());
assert!(encoding::Index::find("BINARY").is_ok());
assert!(encoding::Index::find("none").is_err());