Module magnus::exception

Types and functions for working with Ruby exceptions.

Structs

Exception

Wrapper type for a Value known to be an instance of Ruby’s Exception class.

ExceptionClass

A Value known to be an instance of Class and subclass of Exception.

Functions

arg_error

Return Ruby’s ArgumentError class.

enc_compat_error

Return Ruby’s Encoding::CompatibilityError class.

encoding_error

Return Ruby’s EncodingError class.

eof_error

Return Ruby’s EOFError class.

exception

Return Ruby’s Exception class.

fatal

Return Ruby’s fatal class.

float_domain_error

Return Ruby’s FloatDomainError class.

frozen_error

Return Ruby’s FrozenError class.

index_error

Return Ruby’s IndexError class.

interrupt

Return Ruby’s Interrupt class.

io_error

Return Ruby’s IOError class.

key_error

Return Ruby’s KeyError class.

load_error

Return Ruby’s LoadError class.

local_jump_error

Return Ruby’s LocalJumpError class.

math_domain_error

Return Ruby’s Math::DomainError class.

name_error

Return Ruby’s NameError class.

no_matching_pattern_error

Return Ruby’s NoMatchingPatternError class.

no_matching_pattern_key_error

Return Ruby’s NoMatchingPatternKeyError class.

no_mem_error

Return Ruby’s NoMemoryError class.

no_method_error

Return Ruby’s NoMethodError class.

not_imp_error

Return Ruby’s NotImpError class.

range_error

Return Ruby’s RangeError class.

regexp_error

Return Ruby’s RegexpError class.

runtime_error

Return Ruby’s RuntimeError class.

script_error

Return Ruby’s ScriptError class.

security_error

Return Ruby’s SecurityError class.

signal

Return Ruby’s SignalException class.

standard_error

Return Ruby’s StandardError class.

stop_iteration

Return Ruby’s StopIteration class.

syntax_error

Return Ruby’s SyntaxError class.

sys_stack_error

Return Ruby’s SystemStackError class.

system_call_error

Return Ruby’s SystemCallError class.

system_exit

Return Ruby’s SystemExit class.

thread_error

Return Ruby’s ThreadError class.

type_error

Return Ruby’s TypeError class.

zero_div_error

Return Ruby’s ZeroDivisionError class.