Module magnus::gc

Functions for working with Ruby’s Garbage Collector.

Functions

adjust_memory_usage

Inform Ruby of external memory usage.

all_stats

Returns all possible key/value pairs for stat as a Ruby Hash.

count

Returns the number of garbage collections that have been run since the start of the process.

disable

Disable automatic GC runs.

enable

Enable automatic GC run.

location

Get the new location of an object.

mark

Mark an Object.

mark_movable

Mark an Object and let Ruby know it is moveable.

mark_slice

Mark multiple Objects.

register_address

Inform Ruby’s garbage collector that valref points to a live Ruby object.

register_mark_object

Registers value to never be garbage collected.

start

Trigger a “full” GC run.

stat

Returns the GC profiling value for key.

unregister_address

Inform Ruby’s garbage collector that valref that was previously registered with register_address no longer points to a live Ruby object.