Struct magnus::Binding

#[repr(transparent)]
pub struct Binding(_);

A Value known to be an instance of Binding.

All Value methods should be available on this type through Deref, but some may be missed by this documentation.

Implementations

impl Binding

pub fn from_value(val: Value) -> Option<Self>

Return Some(Binding) if val is a Binding, None otherwise.

pub fn eval<T, U>(&self, s: T) -> Result<U, Error>where
T: IntoRString,
U: TryConvert,

Evaluate a string of Ruby code within the binding’s context.

Examples

use magnus::{eval, Binding};

let binding = eval::<Binding>("binding").unwrap();
assert_eq!(binding.eval::<_, i64>("1 + 2").unwrap(), 3);

pub fn local_variable_get<N, T>(&self, name: N) -> Result<T, Error>where
N: IntoSymbol,
T: TryConvert,

Get the named local variable from the binding.

Returns Ok(T) if the method returns without error and the return value converts to a T, or returns Err if the local variable does not exist or the conversion fails.

Examples

use magnus::{eval, Binding, Value};

let binding = eval::<Binding>("binding").unwrap();
binding.local_variable_set("a", 1);
assert_eq!(binding.local_variable_get::<_, i64>("a").unwrap(), 1);
assert!(binding.local_variable_get::<_, Value>("b").is_err());

pub fn local_variable_set<N, T>(&self, name: N, val: T)where
N: IntoSymbol,
T: IntoValue,

Set the named local variable in the binding.

Examples

use magnus::{eval, Binding};

let binding = eval::<Binding>("binding").unwrap();
binding.local_variable_set("a", 1);
assert_eq!(binding.local_variable_get::<_, i64>("a").unwrap(), 1);