Trait magnus::ArgList

pub trait ArgList {
    type Output: AsRef<[Value]>;

    fn into_arg_list(self) -> Self::Output;
}

Trait for types that can be used as an arguments list when calling Ruby methods.

Required Associated Types

type Output: AsRef<[Value]>

The type of the arguments list. Must convert to &[Value] with AsRef.

Required Methods

fn into_arg_list(self) -> Self::Output

Convert self into a type that can be used as a Ruby argument list.