Trait magnus::IntoValueFromNative

pub unsafe trait IntoValueFromNative: IntoValue { }

Conversions from Rust types that do not contain Value into Value.

This trait is used as a bound in functions such as RArray::from_vec to prevent accepting heap allocated datastructures containing Value, as it is not safe to store a Value on the heap.

Safety

This trait must not be implemented for types that contain Value.