Trait magnus::typed_data::TypedData

pub unsafe trait TypedDatawhere
    Self: Send + Sized,{
    fn class() -> RClass;
    fn data_type() -> &'static DataType;

    fn class_for(value: &Self) -> RClass { ... }
}

A trait for Rust types that can be used with the rb_data_typed_object_wrap API.

Safety

This trait is unsafe to implement as the fields of DataType returned by TypedData::data_type control low level behaviour that can go very wrong if set incorrectly. Implementing this trait is the only way a DataType can be passed to Ruby and result in safety violations, DataType is otherwise safe (but useless) to create.

The TypedData or wrap macros can help implementing this trait more safely.

Required Methods

fn class() -> RClass

Should return the class for the Ruby object wrapping the Rust type.

This can be overridden on a case by case basis by implementing TypedData::class_for, but the result of this function will always be used in error messages if a value fails to convert to Self.

If using class_for it is advised to have this function return the superclass of those returned by class_for.

Examples

use magnus::{prelude::*, class, define_class, memoize, RClass, TypedData};

struct Example();

unsafe impl TypedData for Example {
    fn class() -> RClass {
        *memoize!(RClass: {
          let class = define_class("Example", class::object()).unwrap();
          class.undef_alloc_func();
          class
        })
    }

    // ...
}

fn data_type() -> &'static DataType

Should return a static reference to a DataType with metadata about the wrapped type.

Examples

use magnus::{memoize, typed_data::DataTypeBuilder, DataType, DataTypeFunctions, TypedData};

#[derive(DataTypeFunctions)]
struct Example();

unsafe impl TypedData for Example {
    // ...

    fn data_type() -> &'static DataType {
        memoize!(DataType: DataTypeBuilder::<Example>::new("example").build())
    }
}

Provided Methods

fn class_for(value: &Self) -> RClass

Used to customise the class wrapping a specific value of Self.

The provided implementation simply returns the value of TypedData::class.

Be aware TypedData::class, will always be used in error messages if a value fails to convert to Self. It is advised to have TypedData::class return the superclass of those returned by this function.

Examples

use magnus::{prelude::*, define_class, memoize, RClass, TypedData};

enum Example {
    A,
    B,
}

unsafe impl TypedData for Example {
    // ...

    fn class_for(value: &Self) -> RClass {
        match value {
            Self::A => *memoize!(RClass: {
                let class = define_class("A", <Self as TypedData>::class()).unwrap();
                class.undef_alloc_func();
                class
            }),
            Self::B => *memoize!(RClass: {
                let class = define_class("B", <Self as TypedData>::class()).unwrap();
                class.undef_alloc_func();
                class
            }),
        }
    }
}