Module magnus::value

Types for working with Ruby’s VALUE type, representing all objects, and ‘immediate’ values such as Fixnum.

Structs

BoxValue

Protects a Ruby Value from the garbage collector.

Fixnum

A Value known to be a fixnum, Ruby’s internal representation of small integers.

Flonum

A Value known to be a flonum, Ruby’s internal representation of lower precision floating point numbers.

Id

The internal value of a Ruby symbol.

Qfalse

Ruby’s false value.

Qnil

Ruby’s nil value.

Qtrue

Ruby’s true value.

Qundef

A placeholder value that represents an undefined value. Not exposed to Ruby level code.

StaticSymbol

A static Ruby symbol that will live for the life of the program and never be garbage collected.

Value

Ruby’s VALUE type, which can represent any Ruby object.

Constants

QFALSE

Ruby’s false value.

QNIL

Ruby’s nil value.

QTRUE

Ruby’s true value.

QUNDEF

A placeholder value that represents an undefined value. Not exposed to Ruby level code.

Traits

IntoId

Conversions from Rust types into Id.

ReprValue

Marker trait for types that have the same representation as Value.